requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
requests==2.28.2
python-dotenv==1.0.0